+++ vitrina/datasets/views.py
@@ -43,9 +43,8 @@ from parler.views import (
     LanguageChoiceMixin,
     ViewUrlMixin,
 )
-from reversion import set_comment, create_revision, set_user, add_to_revision
+from reversion import add_to_revision
 from reversion.models import Version
-from reversion.views import RevisionMixin
 
 from vitrina.structure.models import Version as _Version
 from vitrina.api.helpers import get_datasets_for_rdf
@@ -604,7 +603,7 @@ class DatasetDetailView(
         return dataset.get_json_ld(f"{SPINTA_SERVER_URL}/{model}")
 
 
-class DatasetDeleteView(PermissionRequiredMixin, RevisionMixin, DeleteView):
+class DatasetDeleteView(PermissionRequiredMixin, DeleteView):
     model = Dataset
     template_name = "confirm_delete.html"
     success_url = reverse_lazy("dataset-list")
@@ -687,7 +686,6 @@ class DatasetCreateView(
     DatasetBreadcrumbsMixin,
     LoginRequiredMixin,
     PermissionRequiredMixin,
-    RevisionMixin,
     TranslatableCreateView,
     LanguageChoiceMixin,
     PlanMixin,
@@ -864,7 +862,6 @@ class DatasetCreateView(
         tags = form.cleaned_data.get("tags")
         self.object.tags.set(tags)
         self.object.save()
-        set_comment(Dataset.CREATED)
         if not form.cleaned_data.get("creator"):
             Representative.objects.create(
                 content_type=ContentType.objects.get_for_model(self.object),
@@ -993,7 +990,6 @@ class ResourceSubclassCreateView(
     DatasetBreadcrumbsMixin,
     LoginRequiredMixin,
     PermissionRequiredMixin,
-    RevisionMixin,
     CreateView,
 ):
     model = Dataset
@@ -1081,7 +1077,6 @@ class DatasetUpdateView(
     HistoryView,
     LoginRequiredMixin,
     PermissionRequiredMixin,
-    RevisionMixin,
     TranslatableUpdateView,
     ViewUrlMixin,
 ):
@@ -1205,7 +1200,6 @@ class DatasetUpdateView(
             self.object.service_type.set(form.cleaned_data["service_type"])
 
         self.object.save()
-        set_comment(Dataset.EDITED)
 
         if "files" in form.changed_data:
             for file in form.cleaned_data.get("files", []):
@@ -1483,7 +1477,6 @@ class DatasetStructureImportView(
     PlanMixin,
     LoginRequiredMixin,
     PermissionRequiredMixin,
-    RevisionMixin,
     CreateView,
 ):
     model = DatasetStructure
@@ -1531,8 +1524,6 @@ class DatasetStructureImportView(
         self.object.save()
         self.object.dataset.current_structure = self.object
         self.object.dataset.save()
-        set_comment(_(f'Added Structure file "{self.object.file}".'))
-
         if self.metadata_version and self.metadata_version.status != VersionStatus.DRAFT:
             form.add_error("file", _("Negalima importuoti struktūros, kai versijos būsena nėra juodraštis."))
             return self.form_invalid(form)
@@ -2062,7 +2053,6 @@ class DatasetRequestsView(DatasetStructureMixin, PermissionRequiredMixin, Histor
 class AddRequestView(
     LoginRequiredMixin,
     PermissionRequiredMixin,
-    RevisionMixin,
     UpdateView,
 ):
     model = Dataset
@@ -2109,7 +2099,6 @@ class AddRequestView(
             status=Task.CREATED,
             type=Task.REQUEST,
         )
-        set_comment(Dataset.REQUEST_SET)
         self.object.save()
         return HttpResponseRedirect(reverse("dataset-requests", kwargs={"pk": self.object.pk}))
 
@@ -2162,7 +2151,6 @@ class RemoveRequestView(LoginRequiredMixin, PermissionRequiredMixin, DeleteView)
 class AddProjectView(
     LoginRequiredMixin,
     PermissionRequiredMixin,
-    RevisionMixin,
     UpdateView,
 ):
     model = Dataset
@@ -2195,7 +2183,6 @@ class AddProjectView(
         for project in form.cleaned_data["projects"]:
             temp_proj = get_object_or_404(Project, pk=project.pk)
             temp_proj.datasets.add(self.object)
-        set_comment(Dataset.PROJECT_SET)
         self.object.save()
         return HttpResponseRedirect(reverse("dataset-projects", kwargs={"pk": self.object.pk}))
 
@@ -3129,7 +3116,7 @@ class QuarterStatsView(DatasetListView):
         return context
 
 
-class DatasetCategoryView(PermissionRequiredMixin, RevisionMixin, TemplateView):
+class DatasetCategoryView(PermissionRequiredMixin, TemplateView):
     template_name = "vitrina/datasets/dataset_categories.html"
 
     dataset: Dataset
@@ -3154,7 +3141,6 @@ class DatasetCategoryView(PermissionRequiredMixin, RevisionMixin, TemplateView):
             for category in form.cleaned_data.get("category"):
                 self.dataset.category.add(category)
             self.dataset.save()
-            set_comment(Dataset.CATEGORY_UPDATED)
 
             DatasetExcludedGroups.objects.filter(dataset=self.dataset).delete()
             for group in DatasetGroup.objects.filter(
@@ -3214,7 +3200,7 @@ class FilterCategoryView(LoginRequiredMixin, View):
         return JsonResponse({"categories": category_data})
 
 
-class DatasetAttributionCreateView(PermissionRequiredMixin, RevisionMixin, CreateView):
+class DatasetAttributionCreateView(PermissionRequiredMixin, CreateView):
     model = DatasetAttribution
     form_class = DatasetAttributionForm
     template_name = "vitrina/datasets/attribution_form.html"
@@ -3242,7 +3228,6 @@ class DatasetAttributionCreateView(PermissionRequiredMixin, RevisionMixin, Creat
         self.object: DatasetAttribution = form.save(commit=False)
         self.object.dataset = self.dataset
         self.object.save()
-        set_comment(Dataset.ATTRIBUTION_ADDED)
         return redirect(self.dataset.get_absolute_url())
 
 
@@ -3259,11 +3244,8 @@ class DatasetAttributionDeleteView(PermissionRequiredMixin, DeleteView):
         return has_perm(self.request.user, Action.UPDATE, self.dataset)
 
     def form_valid(self, form: BaseForm) -> HttpResponse:
-        with create_revision():
-            self.object = self.get_object()
-            set_user(self.request.user)
-            add_to_revision(self.object)
-            set_comment(Dataset.ATTRIBUTION_DELETED)
+        self.object = self.get_object()
+        add_to_revision(self.object)
 
         return super().form_valid(form)
 
@@ -3271,7 +3253,7 @@ class DatasetAttributionDeleteView(PermissionRequiredMixin, DeleteView):
         return self.dataset.get_absolute_url()
 
 
-class DatasetRelationCreateView(PermissionRequiredMixin, RevisionMixin, CreateView):
+class DatasetRelationCreateView(PermissionRequiredMixin, CreateView):
     model = DatasetRelation
     form_class = DatasetRelationForm
     template_name = "vitrina/datasets/relation_form.html"
@@ -3321,7 +3303,6 @@ class DatasetRelationCreateView(PermissionRequiredMixin, RevisionMixin, CreateVi
             # need to save to update search index
             self.object.dataset.save()
             self.object.part_of.save()
-            set_comment(Dataset.RELATION_ADDED)
 
         return redirect(self.dataset.get_absolute_url())
 
@@ -3339,11 +3320,8 @@ class DatasetRelationDeleteView(PermissionRequiredMixin, DeleteView):
         return has_perm(self.request.user, Action.UPDATE, self.dataset)
 
     def form_valid(self, form: BaseForm) -> HttpResponse:
-        with create_revision():
-            self.object = self.get_object()
-            set_user(self.request.user)
-            set_comment(Dataset.RELATION_DELETED)
-            add_to_revision(self.object)
+        self.object = self.get_object()
+        add_to_revision(self.object)
         return super().form_valid(form)
 
     def get_success_url(self):
@@ -3396,7 +3374,7 @@ class DatasetPlanView(
         return self.dataset
 
 
-class DatasetCreatePlanView(PermissionRequiredMixin, RevisionMixin, TemplateView):
+class DatasetCreatePlanView(PermissionRequiredMixin, TemplateView):
     template_name = "vitrina/plans/plan_form.html"
 
     dataset: Dataset
@@ -3435,7 +3413,6 @@ class DatasetCreatePlanView(PermissionRequiredMixin, RevisionMixin, TemplateView
             if form_type == "create_form":
                 plan = form.save()
                 PlanDataset.objects.create(plan=plan, dataset=self.dataset)
-                set_comment(_(f'Pridėtas terminas "{plan}". Į terminą įtrauktas duomenų rinkinys "{self.dataset}".'))
 
             else:
                 plan_dataset = form.save(commit=False)
@@ -3443,7 +3420,6 @@ class DatasetCreatePlanView(PermissionRequiredMixin, RevisionMixin, TemplateView
                 plan_dataset.save()
                 plan = plan_dataset.plan
                 plan.save()
-                set_comment(_(f'Į terminą "{plan}" įtrauktas duomenų rinkinys "{self.dataset}".'))
 
             Comment.objects.create(
                 content_type=ContentType.objects.get_for_model(self.dataset),
@@ -3469,7 +3445,7 @@ class DatasetCreatePlanView(PermissionRequiredMixin, RevisionMixin, TemplateView
             return render(request=request, template_name=self.template_name, context=context)
 
 
-class DatasetDeletePlanView(PermissionRequiredMixin, RevisionMixin, DeleteView):
+class DatasetDeletePlanView(PermissionRequiredMixin, DeleteView):
     model = PlanDataset
     template_name = "confirm_delete.html"
 
@@ -3484,7 +3460,6 @@ class DatasetDeletePlanView(PermissionRequiredMixin, RevisionMixin, DeleteView):
         self.object.delete()
 
         plan.save()
-        set_comment(_(f'Iš termino "{plan}" pašalintas duomenų rinkinys "{dataset}".'))
 
         if dataset.is_public and dataset.status == Dataset.PLANNED and not dataset.plandataset_set.exists():
             dataset.status = Dataset.INVENTORED
@@ -3522,7 +3497,6 @@ class DatasetDeletePlanDetailView(DatasetDeletePlanView):
         self.object.delete()
 
         plan.save()
-        set_comment(_(f'Iš termino "{plan}" pašalintas duomenų rinkinys "{dataset}".'))
 
         if dataset.is_public and dataset.status == Dataset.PLANNED and not dataset.plandataset_set.exists():
             dataset.status = Dataset.INVENTORED